pyproject.toml
[tool.poetry]
name = "networkstorage"
version = "0.1.0"
description = "API for network storage service"
authors = ["apolumestny <anpolumestny@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.27"
uvicorn = "^0.27.1"
fastapi = "^0.109.2"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
psycopg = "^3.1.18"
psycopg-binary = "^3.1.18"
ipaddresses = "^0.0.2"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sentry-sdk = {extras = ["fastapi"], version = "^1.43.0"}


[tool.poetry.group.test.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
